+++ crypto/math/src/fft/mod.rs
@@ -4,8 +4,6 @@ pub mod bowers_fft;
 pub mod errors;
 pub mod fft;
 #[cfg(feature = "alloc")]
-pub mod polynomial;
-#[cfg(feature = "alloc")]
 pub mod roots_of_unity;
 
 #[cfg(all(test, feature = "alloc"))]
